chore(extractor): port the data extractor to Rust

+++ tfhe-benchmark/benches/high_level_api/bitonic_shuffle.rs
@@ -224,7 +224,7 @@ fn bench_shuffle_config<T>(
         params_name,
         "bitonic_shuffle",
         &OperatorType::Atomic,
-        value_bits,
+        value_bits as u64,
         vec![],
     );
 }
